factor out register implementation of wp exporter class to declutter plugin
