+++ src/Wordpress/Service/WPPrivacyPersonalDataExporter.php
@@ -7,7 +7,37 @@ use CommonsBooking\Helper\Helper;
 /**
  * WP specific implementations for personal data export and deletion
  */
-class WPPersonalDataBookingService {
+class WPPrivacyPersonalDataExporter {
+
+	/**
+	 * Registers all user data exporters ({@link https://developer.wordpress.org/plugins/privacy/adding-the-personal-data-exporter-to-your-plugin/}).
+	 *
+	 * @param array $exporters
+	 *
+	 * @return mixed
+	 */
+	public static function registerUserDataExporters( $exporters ) {
+		$exporters[ COMMONSBOOKING_PLUGIN_SLUG ] = array(
+			'exporter_friendly_name' => __( 'CommonsBooking Bookings', 'commonsbooking' ),
+			'callback'               => array( self::class, 'exportUserBookingsByEmail' ),
+		);
+		return $exporters;
+	}
+
+	/**
+	 * Registers all user data erasers ({@link https://developer.wordpress.org/plugins/privacy/adding-the-personal-data-eraser-to-your-plugin/}).
+	 *
+	 * @param $erasers
+	 *
+	 * @return mixed
+	 */
+	public static function registerUserDataErasers( $erasers ) {
+		$erasers[ COMMONSBOOKING_PLUGIN_SLUG ] = array(
+			'eraser_friendly_name' => __( 'CommonsBooking Bookings', 'commonsbooking' ),
+			'callback'             => array( self::class, 'removeUserBookingsByEmail' ),
+		);
+		return $erasers;
+	}
 
 	/**
 	 * Export user bookings using the supplied email. This is for integration with the WordPress personal data exporter.
